Don't fallback to the wrapper in case there is no candidate
Wrapper might have invalid error code and threadDumps will be empty. The reason why there is no candidate is that first detection happens before the process is started so we need to retry (done by throwing Exception).

OSHI is not needed here since the problem was a race between pid detection and process start and was not related to reliability of `descendants` method.

GitOrigin-RevId: 4a0548247c976d11c9d876739ef6d4c6404d3301