Move all functionality from JUnit5StarterAssistant except for killing processes into into JUnit5Listener.
Rename JUnit5MixedModeListener to JUnit5Listener because it's unclear what MixedMode means.
Move JUnit5MixedModeListener to the intellij.tools.ide.starter.junit5 module from starter.extended.
Rename JUnit5StarterAssistant to KillOutdatedProcesses.

Reasons:
JUnit5MixedModeListener was a thing that launched DevServer, but this is no longer the case, and essentially it is now responsible for killing perTestSupervisorScope and testSuiteSupervisorScope.

perTestSupervisorScope and testSuiteSupervisorScope are currently used in starter, but there is nothing external available to kill these scopes.

If all necessary functionality will be in JUnit5MixedModeListener, then there is no need to use JUnit5StarterAssistant and JUnit5StarterAssistantExtended (essentially a thing for the aggregator) annotations, as JUnit5MixedModeListener is connected through JUnit magic everywhere there is a starter.

GitOrigin-RevId: 12d84235daf5b35f21b0f60153b2b079f5e51c0a