Do not put `DISPLAY` enviornment into host context
At this point host is expected to be already running, because it is responsible to start xvfb for us, so no need to put `DISPLAY` option into start options. If this is a hacky way to pass it somewhere - well, it is a bad way.

The only thing I can see we are accessing it is ffmpeg recording start in `com.intellij.ide.starter.driver.engine.remoteDev.XorgWindowManagerHandler.subscribeToStartRecording`, but it is invoked for the client only and not for the host. At least - it shouldn't.

GitOrigin-RevId: cb1a9e90e203c16e30a8bdec8b86af3348383a94