Explicitly start fluxbox and ffmpeg
We can't use EventBus for starting fluxbox/ffmpeg, because there is no clear way to unsubscribe from event. And if we have a multiple runs of host/client, even if we subscribe for the first client, we are going to be triggered for the next host run in the same test.

We could hack this by checking some client-related stuff in the run context, but feels like it would be better to start things explicitly instead.

GitOrigin-RevId: 204f1566b195ee353389bafa788b76a366f5862e