Make sure that we destroy process in case graceful exit didn't work out
Otherwise if closing takes too much time, client is not closed before the next test and the next test fails. Also if client is not closed gracefully, backend won't be stopped.
We also need to store frontend ProcessHandler to await till client is closed and not to await that backend will be closed when client is closed.

GitOrigin-RevId: d897d8989a3c789ebbb1f0ea2beeaca125c84f82