[build scripts] refactoring: pass platform prefix for the base IDE of JetBrains Client explicitly (IJPL-185957)
Currently, standalone frontend distributions for all IDEs have platformPrefix 'JetBrainsClient'. To distinguish different distributions of JetBrainsClient inside the dev build, the platform prefix of the base IDE was implicitly prepended to the 'platformPrefix' property. This may cause problems with code which uses 'platformPrefix' and expects the proper value, so now an explicit 'baseIdePlatformPrefixForFrontend' property is added.

GitOrigin-RevId: 9f8f429bc02f25d7a0858c74ddff3c0956cce908